+++ gradio/chat_interface.py
@@ -9,7 +9,6 @@ import inspect
 from typing import AsyncGenerator, Callable, Literal, Union, cast
 
 import anyio
-from gradio_client import utils as client_utils
 from gradio_client.documentation import document
 
 from gradio.blocks import Blocks
@@ -29,7 +28,7 @@ from gradio.helpers import special_args
 from gradio.layouts import Accordion, Group, Row
 from gradio.routes import Request
 from gradio.themes import ThemeClass as Theme
-from gradio.utils import SyncToAsyncIterator, async_iteration
+from gradio.utils import SyncToAsyncIterator, async_iteration, async_lambda
 
 
 @document()
@@ -63,7 +62,7 @@ class ChatInterface(Blocks):
         additional_inputs_accordion_name: str | None = None,
         additional_inputs_accordion: str | Accordion | None = None,
         examples: list[str] | list[dict[str, str | list]] | None = None,
-        cache_examples: bool | None = None,
+        cache_examples: bool | Literal["lazy"] | None = None,
         title: str | None = None,
         description: str | None = None,
         theme: Theme | str | None = None,
@@ -129,12 +128,10 @@ class ChatInterface(Blocks):
         self.is_generator = inspect.isgeneratorfunction(
             self.fn
         ) or inspect.isasyncgenfunction(self.fn)
-        self.examples = examples
-        if self.space_id and cache_examples is None:
-            self.cache_examples = True
-        else:
-            self.cache_examples = cache_examples or False
         self.buttons: list[Button | None] = []
+
+        self.examples = examples
+        self.cache_examples = cache_examples
 
         if additional_inputs:
             if not isinstance(additional_inputs, list):
@@ -284,6 +281,8 @@ class ChatInterface(Blocks):
                     inputs=[self.textbox] + self.additional_inputs,
                     outputs=self.chatbot,
                     fn=examples_fn,
+                    cache_examples=self.cache_examples,
+                    _defer_caching=True,
                 )
 
             any_unrendered_inputs = any(
@@ -296,8 +295,8 @@ class ChatInterface(Blocks):
                             input_component.render()
 
             # The example caching must happen after the input components have rendered
-            if cache_examples:
-                client_utils.synchronize_async(self.examples_handler.cache)
+            if examples:
+                self.examples_handler._start_caching()
 
             self.saved_input = State()
             self.chatbot_state = (
@@ -378,7 +377,7 @@ class ChatInterface(Blocks):
                 show_api=False,
                 queue=False,
             ).then(
-                lambda x: x,
+                async_lambda(lambda x: x),
                 [self.saved_input],
                 [self.textbox],
                 show_api=False,
@@ -387,7 +386,7 @@ class ChatInterface(Blocks):
 
         if self.clear_btn:
             self.clear_btn.click(
-                lambda: ([], [], None),
+                async_lambda(lambda: ([], [], None)),
                 None,
                 [self.chatbot, self.chatbot_state, self.saved_input],
                 queue=False,
@@ -401,9 +400,11 @@ class ChatInterface(Blocks):
             if self.submit_btn:
                 for event_trigger in event_triggers:
                     event_trigger(
-                        lambda: (
-                            Button(visible=False),
-                            Button(visible=True),
+                        async_lambda(
+                            lambda: (
+                                Button(visible=False),
+                                Button(visible=True),
+                            )
                         ),
                         None,
                         [self.submit_btn, self.stop_btn],
@@ -411,7 +412,7 @@ class ChatInterface(Blocks):
                         queue=False,
                     )
                 event_to_cancel.then(
-                    lambda: (Button(visible=True), Button(visible=False)),
+                    async_lambda(lambda: (Button(visible=True), Button(visible=False))),
                     None,
                     [self.submit_btn, self.stop_btn],
                     show_api=False,
@@ -420,14 +421,14 @@ class ChatInterface(Blocks):
             else:
                 for event_trigger in event_triggers:
                     event_trigger(
-                        lambda: Button(visible=True),
+                        async_lambda(lambda: Button(visible=True)),
                         None,
                         [self.stop_btn],
                         show_api=False,
                         queue=False,
                     )
                 event_to_cancel.then(
-                    lambda: Button(visible=False),
+                    async_lambda(lambda: Button(visible=False)),
                     None,
                     [self.stop_btn],
                     show_api=False,
@@ -471,7 +472,7 @@ class ChatInterface(Blocks):
         if message["text"] is not None and isinstance(message["text"], str):
             history.append([message["text"], response])
 
-    def _display_input(
+    async def _display_input(
         self, message: str | dict[str, list], history: list[list[str | tuple | None]]
     ) -> tuple[list[list[str | tuple | None]], list[list[str | tuple | None]]]:
         if self.multimodal and isinstance(message, dict):
@@ -631,7 +632,7 @@ class ChatInterface(Blocks):
         async for response in generator:
             yield [[message, response]]
 
-    def _delete_prev_fn(
+    async def _delete_prev_fn(
         self,
         message: str | dict[str, list],
         history: list[list[str | tuple | None]],