+++ js/dataset/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @gradio/dataset
 
+## 0.1.29
+
+### Dependency updates
+
+- @gradio/upload@0.8.3
+- @gradio/client@0.15.0
+
 ## 0.1.28
 
 ### Dependency updates