+++ js/upload/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @gradio/upload
 
+## 0.8.3
+
+### Dependency updates
+
+- @gradio/upload@0.8.3
+- @gradio/client@0.15.0
+
 ## 0.8.2
 
 ### Fixes